move class to the package where it belongs to (#376)
* move class to the package where it belongs to
